+++ docs/changelog.rst
@@ -2,10 +2,12 @@ v1.1 ( to be released)
 =========================
 1. Updated README file (by `Sergio Pereira <mailto:spereira@zscaler.com>`_)
    - Updated instructions of zpa usage
+   - Added method to create access policy
+
 2. Updated zpa_talker for SAML method (by `Rohit Luthra <mailto:rluthra@zscaler.com>`_)
     - Added method for SAML Attribute pull from the customer portal
-3. Updated zpa_talker for Getting Global policy ID (by `Rohit Luthra <mailto:rluthra@zscaler.com>`_)
-    - Added method for fetching the global policy
+    - Added method for fetching the global policy Global policy ID
+
 
 v1.0 ( June 2021)
 =========================